consolidate test utils + linters
