+++ qbraid_qir/qasm3/elements.py
@@ -17,13 +17,7 @@ import uuid
 from abc import ABCMeta, abstractmethod
 from typing import List, Optional, Tuple
 
-from openqasm3.ast import (
-    BitType,
-    ClassicalDeclaration,
-    Program,
-    QubitDeclaration,
-    Statement,
-)
+from openqasm3.ast import BitType, ClassicalDeclaration, Program, QubitDeclaration, Statement
 from pyqir import Context, Module
 
 
@@ -82,9 +76,7 @@ class Qasm3Module:
         elements (List[Statement]): List of openqasm3 Statements.
     """
 
-    def __init__(
-        self, name: str, module: Module, num_qubits: int, num_clbits: int, elements
-    ):
+    def __init__(self, name: str, module: Module, num_qubits: int, num_clbits: int, elements):
         self._name = name
         self._module = module
         self._num_qubits = num_qubits
@@ -132,9 +124,7 @@ class Qasm3Module:
                 statement.type, BitType
             ):
                 name = statement.identifier.name
-                size = (
-                    None if statement.type.size is None else statement.type.size.value
-                )
+                size = None if statement.type.size is None else statement.type.size.value
                 num_clbits += 1 if size is None else size
                 elements.append(_Register((name, size), False))
             else:
